Support keyed files
Adds support for keyed files as well as more flexible and robust
searching.

Now you can have “foo”->”bar” in the theme and search all bundles for
“foo” of type “dat”, in addition to falling back to the main bundle. If
key “foo” is not found, “foo” is assumed to be the file name.

Still supports files of keys with the type appended, e.g. “foo.dat” as
well as the exiting methods (without type) for backwards compatibility.